requirements.txt
Django==3.1.1
Whoosh==2.7.4
asgiref==3.2.10
boto==2.49.0
boto3==1.14.23
botocore==1.17.23
certifi==2020.6.20
chardet==3.0.4
djangorestframework==3.11.1
docutils==0.15.2
gensim==3.8.3
idna==2.10
jieba==0.42.1
jmespath==0.10.0
numpy==1.19.0
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.1
setuptools==40.6.2
six==1.15.0
smart_open==2.1.0
sqlparse==0.3.1
urllib3==1.25.9
whoosh==2.7.4
